# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fossil Museum API
  version: 1.0.0
  description: API REST del sistema de fósiles.
servers:
  - url: http://localhost:4000
paths:
...
  /api/usuarios:
    get:
      summary: Listar usuarios (admin)
      security:
        - bearerAuth: []
      responses:
        "200": { description: Lista de usuarios }
    post:
      summary: Crear usuario (admin)
      security:
        - bearerAuth: []
      responses:
        "200": { description: Usuario creado }
  /api/usuarios/catalogo/roles:
    get:
      summary: Catálogo de roles (admin)
      security:
        - bearerAuth: []
      responses:
        "200": { description: Lista de roles }
  /api/usuarios/{id}:
    get:
      summary: Obtener usuario por ID (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Usuario }
    put:
      summary: Actualizar usuario (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Usuario actualizado }
    delete:
      summary: Soft-delete usuario (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Usuario eliminado }
  /api/usuarios/{id}/roles:
    put:
      summary: Reasignar roles de usuario (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Roles actualizados }
  /api/usuarios/{id}/activo:
    patch:
      summary: Activar/desactivar usuario (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Estado actualizado }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT